### 2016-09-16 #64 @whedon
[REVIEW]: ecopath_matlab: A Matlab-based implementation of the Ecopath food web algorithm
**Submitting author:** kakearney (<a href="http://orcid.org/0000-0002-6152-5236">Kelly Kearney</a>)
**Repository:** <a href="https://github.com/kakearney/ecopath_matlab-pkg" target ="_blank">https://github.com/kakearney/ecopath_matlab-pkg</a>
**Version:** v2.0
