### 2016-09-17 #64 @rmflight
I'm curious if a software package that depends on a *licensed* non-free product counts as **open source software** and is within the scope of the journal?
