### 2016-09-17 #64 @rmflight
I apologize, that was an oversight on my part. I should have looked at the
reviewer guidelines. Thank you for drawing my attention to that.

On Fri, Sep 16, 2016 at 9:34 PM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @rmflight <https://github.com/rmflight> yes code requiring a proprietary
> run-time may be eligible. See also the discussion we had on this here:
> openjournals/joss#142 <https://github.com/openjournals/joss/issues/142>.
> Following that discussion we now say the following here
> http://joss.theoj.org/about:
>
> *What about submissions that rely upon proprietary languages/development
> environments?*
> We strongly prefer software that doesn't rely upon proprietary (paid for)
> development environments/programming languages. However, provided your
> submission meets our submission requirements (including having a valid open
> source license) then we will consider your submission for review. Should
> your submission be accepted for review, we may ask you, the submitting
> author, to help us find reviewers who already have the required development
> environment installed.
>
> And at the reviewer guidelines:
>
> *What about submissions that rely upon proprietary languages/development
> environments?*
> As outlined in our author guidelines, submissions that rely upon a
> proprietary/closed source language or development environment are
> acceptable provided that they meet the other submission requirements and
> that you, the reviewer, are able to install the software & verify the
> functionality of the submission as required by our reviewer guidelines.
> If an open source or free variant of the programming language exists, feel
> free to encourage the submitting author to consider making their software
> compatible with the open source/free variant.
>
