### 2016-09-21 #64 @arfonsmith
Please direct them to volunteer here

> On Sep 21, 2016, at 12:27, Kelly Kearney <notifications@github.com> wrote:
>
> What is the journal's process for suggesting reviewers? Should I compile some suggested names and send them to you (the editor), or should I try to personally contact people and direct them to here to volunteer?
>
