### 2016-09-23 #64 @tpoisot
If it runs in Octave, yes.
