### 2016-09-23 #64 @kakearney
I've never worked with Octave myself, so I'll have to download it and run
some tests.  I'll update the paper accordingly once I've had a chance to do
that.

-Kelly

On Fri, Sep 23, 2016 at 1:56 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @tpoisot <https://github.com/tpoisot> Thanks for your quick response.
> I've asked the authors to clarify compatibility with Octave. I have just
> assigned @roliveros-ramos <https://github.com/roliveros-ramos> as
> reviewer 1. For the moment the form at the top here handles a single
> reviewer (we are expanding our system to easily handle multiple reviewers).
> If you want to review it as well (if compatible with Octave) that would be
> great. For you to contribute as a second reviewer you can leave comments
> here or send me a report via kevin.moerman at gmail.com. Thanks!
>
