### 2016-09-23 #64 @Kevin-Mattheus-Moerman
@kakearney It is likely that it is not fully compatible. A basic statement saying that compatibility has not been confirmed could suffice to. But to do a very quick check for compatibility here are some tips. Identify the most complex MATLAB functions in your submission, or functions requiring a special toolbox, or functions that seem very new in MATLAB (I sometimes, somewhat unfairly, refer to Octave as MATLAB 2005). Then you can check if these functions are part of the default Octave distribution. I ask about adding statements on Octave since the open source community in some cases favors non-proprietary languages like Octave.
