### 2016-10-07 bbcbbc9
Updates to readme
Added note regarding lack of Octave compatibility (a very quick test
showed numerous incompatibilities, including differences in
object-oriented syntax as well as new-ish options in Matlab functions
like sortrows.m).

Also corrected typo in the path setup example (missing parentheses).
