### 2016-10-07 #64 @kakearney
I added a short statement to the README.md file clarifying that this code
will not run in Octave... Octave doesn't support the modern-Matlab class
syntax, along with numerous smaller incompatibilities.

On Thu, Oct 6, 2016 at 7:35 AM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @roliveros-ramos <https://github.com/roliveros-ramos> thanks again for
> agreeing to review this submission. Let me know if you have any questions
> on the review process. Feel free to leave comments to the authors here. Do
> you think you can process the review within 2 weeks?
>
> @kakearney <https://github.com/kakearney> any developments on checking
> compatibility with Octave? Or could you add a statement that is it likely
> not and has not been checked.
>
