### 2016-10-14 #64 @kakearney
There are a few key Matlab features that this software relies on that limit its back-compatibility:
- class syntax: introduced in R2008a
- table arrays: introduced in R2013b
- graph/digraph objects: introduced in R2015b

The last of those will only limit the use of the ecopathmodel graph method, which is really a bonus feature, not necessary to the core calculations.  So I believe this code should be mostly functional under R2014a/b.  It will not work properly in R2012.

As I note at the end of the README, this class-based, table-based version of the code is relatively new.  I decided a reliance on newer Matlab features was worth it to create more user-friendly (and input validation-friendly) storage of the hundreds of parameters in an Ecopath model.  In general, I'm  a big proponent of maintaining back-compatibility, and so I did not make the decision to switch to this format lightly; the use of table arrays streamlined both use and maintenance of the code, which had gotten increasingly difficult as I developed this code from a limited, project-specific implementation to a full Ecopath replication useable by others.

Please note that the old version is still available, and the primary ecopath calculation in that code set (ecopathlite.m) uses nearly identical code to the ecopathmodel ecopath method in this one.  That version can be run in Matlab as old as 2007a, I believe (I make ample use of the bsxfun function, which was introduced at that time).  However, I am not actively maintaining that older version anymore.
