### 2016-11-14 0d7e25d
Corrected Rpath reading function so it can handle models with empty stanza files
