### 2016-11-17 cb50cb1
Added JOSS under review badge to readme
