### 2016-11-25 #64 @roliveros-ramos
@Kevin-Mattheus-Moerman, I'm sorry for the lack of news.
It's been dificult to check the code. I got only MATLAB 2014a for Windows, but I never managed to install the mdbtool library.  I use Fedora but don't have MATLAB for linux available.
I do have C compilers available but I think the installation process for MS Windows is not as simple as described (not described, actually), mainly for the dependencies of mdbtools itself. After several tries I gave up. I checked the code that doesn't need to use mdbtools and it's working fine. I tried to run the rest of the code with other model included in the demo (Esa) and most of it works properly (despite lots of warnings) so I'd guess it should work fine with the proposed model.
As a programmer and ecopath user I've dreamed with a tool like this for a lot of time, but my current opinion is an OS restriction should be added because it's not really functional for Windows, which is a pitty since EwE is provided for Windows only and most of the potential users use Windows. Even if for production use may be use in servers running linux (E2E models?), the added value of reproducibility is still appealing for regular EwE users which relay on Windows for their work.
In summary, I cannot fully answer the section of functionality, and in the current state the installation instructions are not available for Windows users.
Minor stuff: Christensen and Walters 2004 doesn't include a DOI.
