### 2016-11-26 #64 @Kevin-Mattheus-Moerman
@roliveros-ramos thank you so much for your patience and hard work. If you have a MATLAB license associated with your MathWorks account you can likely download the linux version also. Although the download and installation would require quite a lot of your time.... If the authors are able to provide proper instructions for MS Windows, would you be willing to continue the review process at that point? This would be greatly appreciated.

@kakearney can you address @roliveros-ramos 's comments?
Can you consider:
- Fixing the issues for installation under MS Windows
- Adding documentation for installation for MS Windows.

If your software is not meant for MS Windows users please clarify this in the documentation/instructions and the paper.
