### 2016-11-28 a0c29de
Added DOI to Christensen et al 2004 citation in paper.bib
