+++ paper.bib
@@ -14,24 +14,17 @@ year = {2015}
 }
 
 @article{christensen2004,
-abstract = {The Ecopath with Ecosim (EwE) modeling approach combines software for ecosystem trophic mass balance analysis (Ecopath), 
-with a dynamic modeling capability (Ecosim) for exploring past and future impacts of fishing and environmental disturbances 
-as well as for exploring optimal fishing policies. Ecosim models can be replicated over a spatial map grid (Ecospace) to allow 
-exploration of policies such as marine protected areas, while accounting for spatial dispersal/advection effects. 
-The Ecopath approach and software has been under development for two decades, with Ecosim emerging in 1995, and 
-Ecospace in 1998, leading to an integrated and widely applied package. We present an overview of the computational aspects of 
-the Ecopath, Ecosim and Ecospace modules as they are implemented in the most recent software version. The paper summarizes 
-the capabilities of the modeling system with respect to evaluating how fisheries and the environment impact ecosystems. We 
-conclude by a warning about pitfalls in the use of the software for policy exploration. 
-{\{}$\backslash$copyright{\}} 2003 Elsevier B.V. All rights reserved. },
+abstract = {The Ecopath with Ecosim (EwE) modeling approach combines software for ecosystem trophic mass balance analysis (Ecopath), with a dynamic modeling capability (Ecosim) for exploring past and future impacts of fishing and environmental disturbances as well as for exploring optimal fishing policies. Ecosim models can be replicated over a spatial map grid (Ecospace) to allow exploration of policies such as marine protected areas, while accounting for spatial dispersal/advection effects. The Ecopath approach and software has been under development for two decades, with Ecosim emerging in 1995, and Ecospace in 1998, leading to an integrated and widely applied package. We present an overview of the computational aspects of the Ecopath, Ecosim and Ecospace modules as they are implemented in the most recent software version. The paper summarizes the capabilities of the modeling system with respect to evaluating how fisheries and the environment impact ecosystems. We conclude by a warning about pitfalls in the use of the software for policy exploration. {\{}$\backslash$copyright{\}} 2003 Elsevier B.V. All rights reserved.},
 author = {Christensen, Villy and Walters, Carl J},
-file = {:Users/kakearney/Documents/Mendeley Desktop/Christensen, Walters - 2004 - Ecopath with Ecosim methods, capabilities and limitations.pdf:pdf},
+doi = {10.1016/j.ecolmodel.2003.09.003},
+file = {:Users/kakearney/Library/Application Support/Mendeley Desktop/Downloaded/Christensen, Walters - 2004 - Ecopath with Ecosim methods, capabilities and limitations.pdf:pdf},
 journal = {Ecol. Modell.},
 pages = {109--139},
 title = {{Ecopath with Ecosim: methods, capabilities and limitations}},
 volume = {172},
 year = {2004}
 }
+
 
 @article{Steenbeek2015,
 abstract = {Ecopath with Ecosim (EwE) is a widely applied food web model that is mostly known as desktop software for the Microsoft Windows platform. The freely available Microsoft .NET source code of EwE, however, provides a range of possibilities to use the model in different ways, to customize and extend the model, and to execute the model on different operating systems. We provide an overview of the EwE source code, its philosophy, and its technical capabilities, targeted to model builders and advanced users of the EwE software. We showcase novel scientific applications of the EwE model that have been facilitated because of the modular approach of the system. We also present three additional versions of EwE that have been written in different computer languages for dedicated purposes. Lastly, we provide an entry point for users for obtaining and using the source code for custom purposes.},
