### 2016-11-28 #64 @kakearney
Response to reviewer:

> I got only MATLAB 2014a for Windows, but I never managed to install the mdbtool library. I use Fedora but don't have MATLAB for linux available.
I do have C compilers available but I think the installation process for MS Windows is not as simple as described (not described, actually), mainly for the dependencies of mdbtools itself. After several tries I gave up.

The limitation referred to here is related to the mdb2ecopathmodel.m function, one of several tools for importing model data into an ecopathmodel object from other sources.

To clarify, this limitation does not affect the core capabilities of this code package.  The ecopathmodel class is intended to be the focus, providing both data storage (in class object properties) and a variety of Ecopath-related calculations (in the class methods), and the class can be used completely independently of any external compiled software.

That said, I think the ability for a researcher to move data between the different flavors of Ecopath (the original EwE GUI, Rpath, EwE-F, and my ecopath_matlab) without manual typing or copy-pasting is important, and one of the key features that makes my entry stand out from the crowd.  In addition to being able to create and populate an ecopathmodel object manually from Matlab, users can import from EwE-F text files, Rpath-style .csv files, or EwE6 .EwEmdb files.  The latter one is the troublemaker, and unfortunately also the one most users will want to use the most (since in the Ecopath world, 99% of users build their models using the EwE software). And given that, the lack of cross-platform support by the mdb2ecopath.m code is a very frustrating limitation.

A little history on this limitation and why I haven't figured out a universal solution yet... The original Ecopath with Ecosim software was written using Visual Basic; starting with EwEv6, it was ported to Visual .NET.  For data storage, the authors chose to use Microsoft Access database, i.e. .mdb, files.  This is a proprietary, not-publically-documented file format, and there are no native Matlab functions to read this data.  Therefore, I currently rely on a third-party piece of software, known as mdbtools (https://github.com/brianb/mdbtools), which allows one to dump .mdb tables into simple .csv files, and from there I read them into Matlab.  The mdbtools library is well-documented for *nix systems, but not at all for Windows systems (I assume this is because the target audience is those without access to any Microsoft Access drivers, i.e. non-Windows users).  Based on the discussion threads in the mdbtools GitHub repo, it seems that the library can be compiled on Windows.  However, my Windows use is currently limited to a very small virtual machine, and I have not personally accomplished this Windows compilation.  I am currently investigating workarounds for mdb-reading on Windows using the Matlab Database Toolbox; if offers some utilities to use a locally-installed ODBC driver (available only on Windows) for the task.  I should be able to acquire said ODBC driver for my VM shortly, and hope to rectify this limitation in direct read of .Ewemdb files on a Windows OS as soon as possible.  However, even that solution will be pretty limited, since it requires users to also have the ODBC driver (common for Windows users, I think?) plus a separate Matlab Database Toolbox license (not so common toolbox for oceanographers and fisheries scientists, and certainly not free to add).

I'll note that I state pretty clearly in the tutorial (under "Importing Ecopath with Ecosim data") that the lack of MS Windows support for mdb2ecopathmodel.m is a known shortcoming, one I recognize as affecting many users, and one for which I am actively looking for workarounds.  I also provide copies of both example example ecosystems in both EwE6 format (.mdb files) and Rpath .csv format.  So while Windows users may not be able to work through the specific example that imports directly from the .EwEmdb file, they can still build the same ecopathmodel object using alternative import methods (rpath2ecopathmodel.m).

I am currently working on an additional import function to read EwE6 .csv files.  This set of 13 .csv files can be exported directly from the EwE6 software; it's not ideal, since the files need to be exported manually one at a time (and re-exported if the user modifies the model data at all), but it will serve as a stopgap solution so users can get their data from EwE6 to Matlab without needing the mdbtools library, and without having to resort to copying and pasting.  I will post here where that function is completed an added to the repo.

> I tried to run the rest of the code with other model included in the demo (Esa) and most of it works properly (despite lots of warnings) so I'd guess it should work fine with the proposed model.

Could the reviewer clarify what he meant by "most" of the code working?  Was there a particular part of the tutorial (besides direct import) that didn't work, or that provided different answers than expected?

Regarding the warning messages, I note in the tutorial that those messages are to be expected.  One purpose of this rewrite of Ecopath is to provide as much transparency as possible, so any time my code alters a user-provided value, it issues a warning letting the user know.  This includes any changes related to placeholder values.  In the ESA model example, NaNs were used as placeholders for non-relevant parameters (for example, consumption rate of prey for primary producers that don't eat); the code changes these to 0s (to avoid propagation of NaNs in some later calculations), and the warning messages state clearly where this was done and why.

Each of the above warning messages are coded using standard Matlab warning identifiers, and can be silenced by the user if they so choose.

> In summary, I cannot fully answer the section of functionality, and in the current state the installation instructions are not available for Windows users.

I appreciate the attempt to review, but I feel that this statement judges the entire code package based on a rather minor portion of it.  I will work to correct the direct import on Windows limitation, but I believe the ecopathmodel class can stand on its own even without that feature. If necessary, I can recommend some additional reviewers to take a look at the code.

> Minor stuff: Christensen and Walters 2004 doesn't include a DOI.

The DOI for that paper is 10.1016/j.ecolmodel.2003.09.003.  I have added that information to the paper.bib file.
