### 2016-11-29 #64 @kakearney
Thank you for the additional clarification.

As I was preparing to respond to this comment, I received an email from a new ecopath_matlab user asking if I had a workaround for the mdbtools limitation on Windows.  So while I may consider that function a bonus feature, it's now obvious to me that users need a solution to import existing EwE6 models directly on a Windows system.  I will try to get the EwE6 .csv import option working as soon as possible, and will add this option to the tutorial.  I think this should take me about a week.

> The code run but I got errors with most of the plots. Most of them can be related to the 2014a version I'm using (missing histogram function and methods for 'digraph') but I'm not sure

The `histogram` function was introduced in Matlab R2014b (intended to replace `hist` and `histc`), and graph/digraph objects in R2015b.  The histogram plot was simply demonstrating distribution of parameters; I'll go ahead and change that code to use a more back-compatible option.

> but I do understand the full code was not intended to be run with the Esa model.

While the exact verbatim code in the tutorial was designed around several different models (Gen37 represents the simplest sort, Tampa Bay a more complex one with multi-stanza groups and multiple fishing fleets, and Esa a manually-built one with the sort of parameters often included in print journals), all the ecopathmodel methods should work with any set of model data.

> A quick fix can be including Gen37 and Tb as csv files so Windows users can test the core functionality of the package without the mdbtools dependency, but I would stress that it reduces the usability of the package for Windows users (myself have several EwE models in the regular mdb format I cannot inmediately try with your package in the current stage)

The models are currently included in the examples folder in Rpath csv format.  I'll add EwE6 csv format shortly (as well as instructions for exporting models from EwE6 into this format).
