### 2016-12-01 a7bece0
Updated histogram plot in example for backward compatability
