### 2016-12-02 2dd5438
Added warning message related to setting detrital BH after it was zeroed out
