+++ mdbexport/setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+
+setup(
+    name='mdbexport',
+    version='0.1dev',
+    packages=['mdbexport'],
+    author='Kelly Kearney',
+    author_email='kakearney@gmail.com'
+)
