+++ mdbexport/README.md
@@ -6,8 +6,8 @@ This function was written and tested in a Windows environment, and requires the 
 
 It may be possible to get this running on Linux/Mac with the unixODBC driver, but I have not tested that.  I prefer to use the mdbtools library on those platforms, since it already has the unixODBC capabilities bundled.
 
-### Installation instructions
-
-From the command prompt, navigate to this folder ([your-install-location]\ecopath_matlab\mdbexport) and run
+I've added a setup script to this folder, in case any users want to install the module locally.  To do so, from the command prompt, navigate to this folder ([your-install-location]\ecopath_matlab\mdbexport) and run
 
 `python setup.py install`  
+
+However, if you're only using this function via its internal call from mdb2ecopathmodel via Matlab, then the local install isn't necessary.  That function is designed to add this folder to the python system path within Matlab and then read the function directly.
