+++ mdb2ecopathmodel.m
@@ -41,28 +41,58 @@ if ~exist(file, 'file')
     error('Could not find file: %s', file);
 end
 
-% Translate file names such that they can be passed via command line
+% Check that mdb-querying tools are ready to go
 
-file = regexprep(file, '([\s,<>|:\(\)&;\?\*])', '\\$1');
+checkdeps;
 
-% Read table names
+% Dump tables to comma-delimited files
 
-cmd = sprintf('mdb-tables -d, %s', file);
-[s,r] = system(cmd);
-if s
-    if s == 127
-        msg = 'Error calling mdb-tables.  Please make sure you have installed the mdbtools utility and that it is accessible to Matlab (Note that your MATLAB system path, seen with getenv(''PATH''), may differ from your Matlab search path; the former is what matters here)';
-        msg = wraptext(msg, 70);
-        error('%s', msg);
-    else
+csvfolder = tempname;
+if ~exist(csvfolder, 'dir')
+    mkdir(csvfolder);
+end
+
+if ispc
+    
+    py.mdbexport.mdbexport(csvfolder, file);
+    
+    tables = dir(fullfile(csvfolder, '*.csv'));
+    tables = strrep({tables.name}, '.csv', '');
+else
+    % Translate file name such that it can be passed via command line
+
+    file = regexprep(file, '([\s,<>|:\(\)&;\?\*])', '\\$1'); 
+
+    % Read table names
+
+    cmd = sprintf('mdb-tables -1 %s', file); % one per line to account for commas, spaces in names
+    [s,r] = system(cmd);
+    if s
         error('Error reading table names: %s', r);
     end
-end
+
+    tables = regexp(r, '\n', 'split');
+    isemp = cellfun('isempty', tables);
+    tables = tables(~isemp); 
     
-tables = regexp(r, ',', 'split');
-tables = regexprep(tables, '\n', '');
-isemp = cellfun('isempty', tables);
-tables = tables(~isemp);  
+    % Extract to csv
+       
+    for it = 1:length(tables)
+        fname = fullfile(csvfolder, tables{it});
+        if regexpfound(tables{it}, '\s')
+            tbl = regexprep(tables{it}, '([\s,<>|:\(\)&;\?\*])', '\\$1');
+            tables{it} = regexprep(tables{it}, '\s', '_');
+            fname = regexprep(fname, '\s', '_');
+            cmd = sprintf('mdb-export %s %s > %s.csv', file, tbl, fname);
+        else
+            cmd = sprintf('mdb-export %s %s > %s.csv', file, tables{it}, fname);
+        end
+        [s,r] = system(cmd);
+        if s
+            err = [err; {tables{it} r}];
+        end
+    end
+end
 
 % Crazy regular expression for Excel-style comma-delimited stuff... find
 % commas that aren't imbedded within quotes.  No longer needed with
@@ -71,30 +101,28 @@ tables = tables(~isemp);
 
 pattern = ',(?=(?:[^\"]*\"[^\"]*\")*(?![^\"]*\"))';
 
+% Check that this is an EwE6, not EwE5 (or any other random .mdb) file
+
+tbls = {'EcopathGroup', 'EcopathFleet', 'Stanza', 'EcopathDietComp', ...
+        'EcopathCatch', 'EcopathDiscardFate', 'StanzaLifeStage', ...
+        'EcopathGroupPedigree', 'Pedigree'};
+hastbl = ismember(tbls, tables);
+if ~all(hastbl)
+    str = sprintf('%s, ', tbls{~hastbl});
+    error('Expected tables (%s) not found in file; please check that this is an EwE6-formatted file', str(1:end-1));
[…]
