### 2016-12-02 c175392
Corrected typo in error checking, and silenced table name warning
