### 2016-12-03 726816c
Corrected typo in reading text of unparseable tables
