### 2016-12-03 516c464
Added error message to graph method if run in pre-R2015b Matlab
