### 2016-12-03 f4dc306
Minor tweaks to README
