### 2016-12-05 0ba5f3c
Updated ecopathmodel graph error message to use older version syntax
