+++ @ecopathmodel/graph.m
@@ -51,7 +51,7 @@ function G = graph(EM, varargin)
 
 % Copyright 2016 Kelly Kearney
 
-if verLessThan('matlab', 'R2015b')
+if verLessThan('matlab', '8.6.0') % R2015b
     error('The ecopathmodel graph method requires Matlab R2015b or later');
 end
 
