### 2016-12-05 084b47c
Updates to overview document
Added some discussion on the use of the Rpath csv format as a
backup for import of EwE6 data if direct read doesn't work.

Also changed the histogram plot; for some reason the histogram command
doesn't dyamically assign the correct number of bins on a Windows
system when all values are the same (as in the all-1s sum)
