### 2017-01-05 #64 @roliveros-ramos
@Kevin-Mattheus-Moerman, I have finished the review. I really appreciate the efforts from @kakearney  to make it easier to use to Windows users.  The python functions works from MATLAB 2014b, so I changed from 2014a to 2016a, which is the MATLAB version I used to this review. I was able to fully run the demo included with the package and it's working as expected, I still have some comments:

1. `mdb2ecopathmodel` function: I needed to change the `mdbexport.py` (line 33) to:
`DRV = '{Microsoft Access Driver (*.mdb, *.accdb)}'`
because the name of the driver in my computer was not the one provided by default. Without the proper driver name, you get the error:

> python error: [im002] [microsoft][odbc driver manager] datasource name not found and no default driver specified

 I'd be nice to have a warning or some workaround for that.

2. `ecopathmodel2rpath` function. It works ok, but cannot be reverted using the `rpath2ecopathmodel` functions:
`Gen37 = rpath2ecopathmodel(fullfile('examples', 'Gen37', 'gen37'));`
 since more files are required than the ones written by the function (e.g. `Unable to open file 'examples\Gen37\gen37_pedigree.csv'.`). I made it work creating empty "pedigree", "stanzas" and "stanza_groups" (just with the proper headers). The same applied to the TampaBay model.

3. The headers of the file `tb_stanza_groups.csv` are incorrect, `VGFB` need to be changed to `VBGF` to make it work.

One additional comment which may be useful for the installation instructions: MS Office installed by default is the 32 bits one if you already have other 32bits MS Office stuff (any driver or viewer, very likely if you have your computer more than a few years for example), so normally ODBC drivers are 32 bits (or it was my case). If your MATLAB version is 64 bits you need to install Python 64 bits and have 64 bits ODBC drivers. You may have to reinstall some of the programs to make them compatible. I reinstalled MS Office 64bits since we have licenses for that. Not sure if it's the only solution, but was the simpler.
