### 2017-01-06 #64 @Kevin-Mattheus-Moerman
Great work! Thanks a lot @roliveros-ramos ! @kakearney can you please address these additional comments?
