+++ mdbexport/mdbexport/mdbexport.py
@@ -28,12 +28,17 @@ def mdbexport(outdir, mdbfile):
 	if not os.path.isdir(outdir):
 		os.mkdir(outdir)
 
+    # Look for mdb driver
+    
+    DRV = ['{{{}}}'.format(s) for s in pyodbc.drivers() if 'Microsoft Access Driver' in s]
+    if not DRV:
+        raise ValueError('mdbexport did not find a valid Microsoft Access Driver on this computer')
+
 	# Connect to database
 
-	DRV = '{Microsoft Access Driver (*.mdb)}'
 	PWD = 'pw'
 
-	con = pyodbc.connect('DRIVER={};DBQ={};PWD={};CHARSET=UTF8'.format(DRV,mdbfile,PWD))
+	con = pyodbc.connect('DRIVER={};DBQ={};PWD={};CHARSET=UTF8'.format(drv,mdbfile,PWD))            
 	cur = con.cursor()
 
 	# List tables in file, filtering out system tables
