### 2017-01-06 1aac3a2
Corrected a few minor typos in mdb2ecopathmodel help
