### 2017-01-06 f17e680
Updated overview document to be compatible with new rpath2ecopathmodel syntax
