### 2017-01-07 #64 @Kevin-Mattheus-Moerman
@roliveros-ramos can you review how @kakearney addressed your comments? Let me know if you have additional comments or if you feel this submission is now suitable for publication.
