### 2017-01-08 #64 @roliveros-ramos
I just pulled the lasted code and I'm getting this error:

> Error using mdb2ecopathmodel>checkdeps (line 531)
> Could not import mdbexport python module; please check that you either installed it locally or kept it in its default location in the ecopath_matlab package
>
> Error in mdb2ecopathmodel (line 106)
> checkdeps;

Points 2 and 3 are fine!
