### 2017-01-09 d4f37af
Corrected a few typos in mdbexport.py
Corrected a spaces-vs-tab typo, and a list instead of string typo,
both which were introduced in the previous update.
