### 2017-01-09 #64 @kakearney
Sorry, pushed that previous update without testing it; a small syntax error (spaces instead of tabs) prevented Matlab from properly importing the module.  Just corrected that (and tested it on Windows this time)... should be good to go now.
