### 2017-01-12 #64 @roliveros-ramos
@Kevin-Mattheus-Moerman: everything looks fine now, I recommend the paper is accepted for publication.
