### 2017-01-12 #64 @arfon
BTW, here's the latest paper:
[10.21105.joss.00064.pdf](https://github.com/openjournals/joss-reviews/files/702296/10.21105.joss.00064.pdf)
