### 2017-01-12 #64 @kakearney
I just issued a new release, v2.1.  The new DOI is 10.5281/zenodo.17838
