### 2017-01-12 294fecd
WIP: removed DOI badge (wrong one)
