### 2017-01-12 d571538
Added DOI badge to readme (correct one this time)
